+++ packages/dara-components/dara/components/common/switch.py
@@ -15,11 +15,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from typing import Optional, Union
+from typing import Optional
 
 from dara.components.common.base_component import FormComponent
 from dara.core.base_definitions import Action
-from dara.core.interactivity import UrlVariable, Variable
+from dara.core.interactivity import Variable
 
 
 class Switch(FormComponent):
@@ -44,6 +44,6 @@ class Switch(FormComponent):
     :param id: the key to be used if this component is within a form
     """
 
-    value: Optional[Union[Variable[bool], UrlVariable[bool]]] = None
+    value: Optional[Variable[bool]] = None
     onchange: Optional[Action] = None
     id: Optional[str] = None