+++ packages/dara-core/js/shared/interactivity/internal.tsx
@@ -1,8 +1,8 @@
 // ESLint does not understand this pattern, see: https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de
 /* eslint-disable import/no-cycle */
 export * from './plain-variable';
-export * from './url-variable';
 export * from './data-variable';
+export * from './server-variable';
 export * from './derived-variable';
 export * from './nested';
 export * from './triggers';